requirements.txt
**Installationer som krävs:**
streamlit
numpy
joblib
pillow
streamlit-drawable-canvas
scikit-learn
pandas
plotly